Overall refactorization
