+++ src/main/java/com/jdbc/demo/persistence/psql/DriverManager.java
@@ -1,7 +1,7 @@
-package com.jdbc.demo.services;
+package com.jdbc.demo.persistence.psql;
 
 import com.jdbc.demo.DriverDAO;
-import com.jdbc.demo.domain.Driver;
+import com.jdbc.demo.domain.psql.Driver;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.slf4j.Logger;
