+++ src/main/java/com/jdbc/demo/persistence/psql/FreightTransportManager.java
@@ -1,9 +1,9 @@
-package com.jdbc.demo.services;
+package com.jdbc.demo.persistence.psql;
 
 import com.jdbc.demo.FreightTransportDAO;
-import com.jdbc.demo.domain.Driver;
-import com.jdbc.demo.domain.FreightTransport;
-import com.jdbc.demo.domain.Vehicle;
+import com.jdbc.demo.domain.psql.Driver;
+import com.jdbc.demo.domain.psql.FreightTransport;
+import com.jdbc.demo.domain.psql.Vehicle;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.slf4j.Logger;
