+++ src/main/java/com/jdbc/demo/persistence/psql/VehicleManager.java
@@ -1,7 +1,7 @@
-package com.jdbc.demo.services;
+package com.jdbc.demo.persistence.psql;
 
 import com.jdbc.demo.VehicleDAO;
-import com.jdbc.demo.domain.Vehicle;
+import com.jdbc.demo.domain.psql.Vehicle;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.slf4j.Logger;
