+++ src/main/java/com/jdbc/demo/web/mvc/ClientController.java
@@ -1,4 +1,4 @@
-package com.jdbc.demo.web;
+package com.jdbc.demo.web.mvc;
 
 import com.jdbc.demo.ClientDAO;
 
@@ -7,7 +7,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 /**
