+++ src/main/java/com/jdbc/demo/web/mvc/DriverController.java
@@ -1,12 +1,12 @@
-package com.jdbc.demo.web;
+package com.jdbc.demo.web.mvc;
 
 import com.jdbc.demo.AddressDAO;
 import com.jdbc.demo.DriverDAO;
 import com.jdbc.demo.FreightTransportDAO;
-import com.jdbc.demo.domain.Driver;
-import com.jdbc.demo.telemetry.maps.MapsConfiguration;
-import com.jdbc.demo.telemetry.mongo.domain.RouteWaypoint;
-import com.jdbc.demo.telemetry.mongo.services.RouteWaypointRepository;
+import com.jdbc.demo.domain.psql.Driver;
+import com.jdbc.demo.MapsConfiguration;
+import com.jdbc.demo.domain.mongo.RouteWaypoint;
+import com.jdbc.demo.persistence.mongo.RouteWaypointRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
