+++ src/main/java/com/jdbc/demo/web/mvc/IndexController.java
@@ -1,4 +1,4 @@
-package com.jdbc.demo.web;
+package com.jdbc.demo.web.mvc;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
