+++ src/main/java/com/jdbc/demo/web/rest/DriverRestController.java
@@ -1,7 +1,7 @@
-package com.jdbc.demo.web;
+package com.jdbc.demo.web.rest;
 
 import com.jdbc.demo.DriverDAO;
-import com.jdbc.demo.domain.Driver;
+import com.jdbc.demo.domain.psql.Driver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
