+++ src/main/java/com/jdbc/demo/web/rest/VehicleRestController.java
@@ -1,7 +1,7 @@
-package com.jdbc.demo.web;
+package com.jdbc.demo.web.rest;
 
 import com.jdbc.demo.VehicleDAO;
-import com.jdbc.demo.domain.Vehicle;
+import com.jdbc.demo.domain.psql.Vehicle;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
