+++ src/main/webapp/WEB-INF/mvc-dispatcher-servlet.xml
@@ -9,7 +9,7 @@
 
     <context:component-scan base-package="com.jdbc.demo" />
 
-    <bean class="com.jdbc.demo.web.DriverController" />
+    <bean class="com.jdbc.demo.web.mvc.DriverController" />
 
     <bean class="org.springframework.web.servlet.mvc.support.ControllerClassNameHandlerMapping" />
