+++ src/test/java/com/jdbc/demo/services/AddressEntityManagerTest.java
@@ -1,7 +1,7 @@
 package com.jdbc.demo.services;
 
 import com.jdbc.demo.AddressDAO;
-import com.jdbc.demo.domain.Address;
+import com.jdbc.demo.domain.psql.Address;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
