+++ src/test/java/com/jdbc/demo/services/ClientEntityManagerTest.java
@@ -2,8 +2,8 @@ package com.jdbc.demo.services;
 
 import com.jdbc.demo.AddressDAO;
 import com.jdbc.demo.ClientDAO;
-import com.jdbc.demo.domain.Address;
-import com.jdbc.demo.domain.Client;
+import com.jdbc.demo.domain.psql.Address;
+import com.jdbc.demo.domain.psql.Client;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
